week2/airflow/requirements.txt
apache-airflow-providers-google
pyarrow
request